+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.12.0
+
+* Add `StreamQueue.startTransaction()` and `StreamQueue.startTransactions()`.
+  These allow users to conditionally consume events based on their values.
+
 ## 1.11.3
 
 * Fix strong-mode warning against the signature of Future.then